Issue #3204322 by tBKoT: Fixes after review
