+++ modules/social_features/social_profile/config/install/social_profile.settings.yml
@@ -1,5 +1,5 @@
 social_profile_show_email: true
 social_profile_show_language: true
-enable_profile_tagging: true
-allow_category_split: true
-use_category_parent: true
+enable_profile_tagging: false
+allow_category_split: false
+use_category_parent: false
