+++ modules/social_features/social_profile/social_profile.links.menu.yml
@@ -1,5 +1,5 @@
 social_profile.settings:
  title: 'Profile settings'
- description: 'Profile settings for Social Profile module.'
+ description: 'Privacy settings for Social Profile module.'
  route_name: social_profile.settings
  parent: user.admin_index
