+++ modules/social_features/social_profile/src/Plugin/Field/FieldWidget/SocialProfileTagSplitWidget.php
@@ -1,20 +1,26 @@
 <?php
 
-namespace Drupal\social_profile\Plugin\views\filter;
+namespace Drupal\social_profile\Plugin\Field\FieldWidget;
 
+use Drupal\Core\Field\FieldItemListInterface;
 use Drupal\Core\Form\FormStateInterface;
-use Drupal\search_api\Plugin\views\filter\SearchApiTerm;
+use Drupal\select2\Plugin\Field\FieldWidget\Select2EntityReferenceWidget;
 use Drupal\social_profile\SocialProfileTagServiceInterface;
 use Symfony\Component\DependencyInjection\ContainerInterface;
 
 /**
- * Defines a filter for filtering on taxonomy term references.
+ * Class SocialProfileTagSplitWidget.
  *
- * @ingroup views_filter_handlers
- *
- * @ViewsFilter("social_search_api_split_terms")
+ * @FieldWidget (
+ *   id = "social_profile_tag_split",
+ *   label = @Translation("Social Profile Tag Split"),
+ *   field_types = {
+ *     "entity_reference",
+ *   },
+ *   multiple_values = TRUE
+ * )
  */
-class SocialSearchApiSplitTerms extends SearchApiTerm {
+class SocialProfileTagSplitWidget extends Select2EntityReferenceWidget {
 
   /**
    * The social profile tag service.
@@ -27,9 +33,9 @@ class SocialSearchApiSplitTerms extends SearchApiTerm {
    * {@inheritdoc}
    */
   public static function create(ContainerInterface $container, array $configuration, $plugin_id, $plugin_definition) {
-    $filter = parent::create($container, $configuration, $plugin_id, $plugin_definition);
-    $filter->setSocialProfileTagService($container->get('social_profile.tag_service'));
-    return $filter;
+    $widget = parent::create($container, $configuration, $plugin_id, $plugin_definition);
+    $widget->setSocialProfileTagService($container->get('social_profile.tag_service'));
+    return $widget;
   }
 
   /**
@@ -45,15 +51,8 @@ class SocialSearchApiSplitTerms extends SearchApiTerm {
   /**
    * {@inheritdoc}
    */
-  protected function valueForm(&$form, FormStateInterface $form_state) {
-    parent::valueForm($form, $form_state);
-
-    // Use only for profile tag field.
-    if ($this->field !== 'field_profile_profile_tag') {
-      return;
-    }
-
-    $element = &$form['value'];
+  public function formElement(FieldItemListInterface $items, $delta, array $element, array &$form, FormStateInterface $form_state) {
+    $element = parent::formElement($items, $delta, $element, $form, $form_state);
 
     // Do not render field if no options or not active.
     if (
@@ -61,29 +60,31 @@ class SocialSearchApiSplitTerms extends SearchApiTerm {
       !$this->profileTagService->hasContent() ||
       empty($element['#options'])
     ) {
-      $element = [];
+      return [];
+    }
+
+    // Only for fields related to taxonomy terms.
+    if ($element['#target_type'] !== 'taxonomy_term') {
+      return $element;
     }
 
     // Render all value if split tag disabled.
     if (!$this->profileTagService->allowSplit()) {
-      $element['#type'] = 'select2';
       $options = [];
       foreach ($this->profileTagService->getCategories() as $key => $value) {
         $options[$value] = $this->profileTagService->getChildrens($key);
       }
       $element['#options'] = $options;
-      return;
+      return $element;
     }
 
-    // Get selected values.
-    $identifier = $this->options['expose']['identifier'];
-    $default_value = $form_state->getUserInput()[$identifier];
+    // Get default values.
+    $default_value = $element['#default_value'];
 
-    // Wrapper.
     $element = [
       '#type' => 'details',
       '#open' => TRUE,
-      '#title' => $element['#title'],
+      '#element_validate' => [[get_class($this), 'validateElement']],
       '#access' => FALSE,
     ];
 
@@ -106,46 +107,42 @@ class SocialSearchApiSplitTerms extends SearchApiTerm {
           '#type' => 'select2',
           '#title' => $category,
           '#multiple' => TRUE,
-          '#value' => $default_value,
+          '#default_value' => $default_value,
           '#options' => $options,
-          '#name' => 'profile_tag',
         ];
 
         $element['#access'] = TRUE;
       }
     }
+    return $element;
   }
 
   /**
    * {@inheritdoc}
    */
-  public function validateExposed(&$form, FormStateInterface $form_state) {
-    if (
-      $this->field !== 'field_profile_profile_tag' ||
-      !$this->profileTagService->allowSplit()
-    ) {
-      parent::validateExposed($form, $form_state);
-      return;
-    }
-
-    $profile_tag_values = [];
-    $identifier = $this->options['expose']['identifier'];
-
-    // Get the main categories.
-    $categories = $this->profileTagService->getCategories();
-
-    // Get form values.
-    $form_values = $form_state->getValues();
-    foreach ($categories as $tid => $category) {
-      $field_name = 'profile_tagging_' . $this->profileTagService->tagLabelToMachineName($category);
-      if (isset($form_values[$field_name])) {
-        $profile_tag_values += $form_values[$field_name];
-        unset($form_values[$field_name]);
+  public static function validateElement(array $element, FormStateInterface $form_state) {
+    /** @var \Drupal\social_profile\SocialProfileTagServiceInterface $profile_tag_service */
+    $profile_tag_service = \Drupal::service('social_profile.tag_service');
+    if ($profile_tag_service->allowSplit()) {
+      $build_info = $form_state->getBuildInfo();
+      if ($build_info['form_id'] === 'user_register_form') {
+        $field_element = ['entity_profile', $element['#field_name']];
       }
+      else {
+        $field_element = $element['#field_name'];
+      }
+      $value = $form_state->getValue($field_element);
+      $field_value = [];
+      // Get the main categories.
+      $categories = $profile_tag_service->getCategories();
+      foreach ($categories as $tid => $category) {
+        $field_name = 'profile_tagging_' . $profile_tag_service->tagLabelToMachineName($category);
+        if (isset($value[$field_name])) {
+          $field_value += $value[$field_name];
+        }
+      }
+      $form_state->setValue($field_element, $field_value);
     }
-    $form_values[$identifier] = $profile_tag_values;
-    $form_state->setValues($form_values);
-    parent::validateExposed($form, $form_state);
   }
 
 }
